# pkg/api/docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  v1.SystemVersion:
    type: object
  v1.Teardown:
    properties:
      id:
        $ref: '#/definitions/v1.TeardownID'
        description: |
          ID
        type: object
      state:
        $ref: '#/definitions/v1.TeardownState'
        description: |
          State. ["pending", "in progress", "succeeded", "failed"]
        type: object
    type: object
  v1.TeardownID:
    type: object
  v1.TeardownState:
    type: object
host: localhost:8876
info:
  contact: {}
  description: Welcome to lattice api
  license:
    name: Apache 2.0
  termsOfService: TBD
  title: Lattice API
  version: "1.0"
paths:
...
  /systems/{systemId}/teardowns:
    get:
      consumes:
      - application/json
      description: list teardowns
      operationId: list-teardowns
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.Teardown'
            type: array
      summary: Lists teardowns
    post:
      consumes:
      - application/json
      description: teardown system
      operationId: teardown-system
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Teardown'
            type: object
      summary: Teardown system
  /systems/{systemId}/teardowns/{id}:
    get:
      consumes:
      - application/json
      description: get teardown
      operationId: get-teardown
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      - description: Teardown ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Teardown'
            type: object
      summary: Get teardown
  /systems/{systemId}/versions:
    get:
      consumes:
      - application/json
      description: list teardowns
      operationId: list-system-versions
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.SystemVersion'
            type: array
      summary: Lists system versions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apiKey
    type: apiKey
swagger: "2.0"